refactor: backend/scriptsディレクトリを整理
