+++ backend/scripts/maintenance/clear_battle_results.py
