+++ backend/scripts/verify/verify_ace_pilots.py
