+++ backend/scripts/verify/verify_advanced_logic.py
