+++ backend/scripts/verify/verify_terrain_detection.py
